refactor: Redo gulp stuff...
